move result types to one file
